# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Nestling API
  version: "1"
  description: |
    The HTTP API of the Nestling server, for generating client code. `docs/API.md` explains the
    same API in prose; keep both in step with `src/routes/` (`tests/openapi.rs` checks that every
    route is listed here).

    - **Auth:** `Authorization: Bearer <token>`. `EventSource` (which can't set headers) may pass
      `?access_token=<token>` instead.
    - **Times in requests:** `"now"`, a local time in the family's time zone
      (`"2026-10-08T14:30"`) or RFC 3339 with an offset. An omitted `start` means now.
    - **Times in responses:** RFC 3339; family data in the family's time zone, account and token
      data in UTC.
    - **Units:** metric only (`_ml`, `_g`, `_cm`, `_c` = °C, `_seconds`).
    - **IDs:** opaque strings (UUIDv7). Where a client may supply its own id it must be a UUID.
    - Things you don't have access to answer `404`, like things that don't exist.
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
  - url: https://nestling.example.com/api/v1
security:
  - bearer: []
...
paths:
...
  /events/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [events]
      operationId: getEvent
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Event" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [events]
      operationId: updateEvent
      summary: Send only what changes; `null` removes a field
      description: |
        Any field of `EventInput` except `id`. Changing `type` is allowed (fields of the old type
        that don't apply are dropped). `start` can't be removed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
            example: { amount_ml: 150, note: null }
      responses:
        "200":
          description: The updated event
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Event" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [events]
      operationId: deleteEvent
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }

  # ---------------------------------------------------------------- timers
...
  /events/{id}/photo:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [events]
      operationId: getEventPhoto
      summary: The entry's photo (a memory in the baby book)
      responses:
        "200":
          description: The image
          content:
            image/jpeg: { schema: { type: string, format: binary } }
            image/png: { schema: { type: string, format: binary } }
            image/webp: { schema: { type: string, format: binary } }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [events]
      operationId: putEventPhoto
      summary: Set the entry's photo (raw JPEG, PNG or WebP, at most 5 MB)
      description: >-
        One photo per entry; any entry type takes one (the app offers it on milestones). The
        entry's `photo_version` and `updated_at` change, so `GET /families/{id}/sync` brings it.
      requestBody:
        required: true
        content:
          image/jpeg: { schema: { type: string, format: binary } }
          image/png: { schema: { type: string, format: binary } }
          image/webp: { schema: { type: string, format: binary } }
      responses:
        "200":
          description: The event
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Event" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [events]
      operationId: deleteEventPhoto
      summary: Remove the entry's photo
      responses:
        "204": { description: Removed (or there was none) }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /families/{id}/sync:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [sync]
      operationId: pullSync
      summary: Everything, or what changed since a cursor
      parameters:
        - name: since
          in: query
          description: The `cursor` of the previous sync; omit for everything.
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [cursor, full, children, events, timers]
                properties:
                  cursor: { type: integer, format: int64, description: store it for next time }
                  full: { type: boolean, description: true without `since` }
                  children:
                    type: array
                    items: { $ref: "#/components/schemas/Child" }
                  events:
                    type: array
                    description: With `since`, deletions come as `DeletedEvent`.
                    items: { $ref: "#/components/schemas/EventOrDeleted" }
                  timers:
                    type: array
                    items: { $ref: "#/components/schemas/Timer" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [sync]
      operationId: pushSync
      summary: Push changes made offline
      description: |
        Each record is settled on its own. The most recent change wins (a `changed_at` in the
        future counts as now); a deletion is final; a timer started offline while another device
        started the same kind for the same child keeps the earlier start. Pushing the same batch
        twice changes nothing. At most 5000 records per push.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                events:
                  type: array
                  items: { $ref: "#/components/schemas/SyncEventIn" }
                timers:
                  type: array
                  items: { $ref: "#/components/schemas/SyncTimerIn" }
      responses:
        "200":
          description: One result per record
          content:
            application/json:
              schema:
                type: object
                required: [events, timers]
                properties:
                  events:
                    type: array
                    items:
                      allOf:
                        - $ref: "#/components/schemas/SyncResult"
                        - type: object
                          properties:
                            current: { allOf: [{ $ref: "#/components/schemas/EventOrDeleted" }], nullable: true }
                  timers:
                    type: array
                    items:
                      allOf:
                        - $ref: "#/components/schemas/SyncResult"
                        - type: object
                          properties:
                            current: { allOf: [{ $ref: "#/components/schemas/Timer" }], nullable: true }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ---------------------------------------------------------------- import / export
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: Doesn't exist, or you don't have access
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [bad_request, unauthorized, forbidden, not_found, conflict, too_many_requests, upstream_error, internal]
            message: { type: string }
      example: { error: { code: bad_request, message: "a diaper must be wet, dirty or dry" } }
    TimeInput:
      type: string
      description: '"now", a local time in the family time zone ("2026-10-08T14:30") or RFC 3339.'
      example: "2026-10-08T14:30"
...
    Sex:
      type: string
      enum: [female, male, other]
    Side:
      type: string
      enum: [left, right, both]
    TimerKind:
      type: string
      enum: [breastfeed, pump, sleep]

    # -------- accounts
...
    Child:
      type: object
      required: [id, family_id, name, birth_date, sex, created_at, updated_at, photo_version, medicines, reminders, book]
      properties:
        id: { type: string }
        family_id: { type: string }
        name: { type: string }
        birth_date: { type: string, format: date, nullable: true }
        sex: { allOf: [{ $ref: "#/components/schemas/Sex" }], nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        photo_version:
          type: integer
          format: int64
          nullable: true
          description: Changes whenever the photo does; null without a photo.
        medicines:
          type: array
          items: { $ref: "#/components/schemas/Medicine" }
        reminders:
          type: array
          items: { $ref: "#/components/schemas/Reminder" }
        book: { $ref: "#/components/schemas/Book" }
    Book:
      type: object
      description: >-
        The baby book's own pages as short texts by key (the app uses e.g. `birth_time`,
        `birth_place`, `hair`, `eyes`, `full_name`, `name_why`, `world_songs`, `price_coffee`).
        Keys: 1–40 lowercase letters, digits or `_`; at most 80 of them; values are strings of at
        most 4,000 characters. Empty strings and nulls are dropped.
      additionalProperties: { type: string, maxLength: 4000 }
    Medicine:
      type: object
      description: |
        A medicine schedule. Its doses are the child's `health` events with `kind: medicine` and the
        same name (case and extra spaces ignored).
      required: [name, every_hours]
      properties:
        name: { type: string, maxLength: 80 }
        every_hours: { type: number, minimum: 0.5, maximum: 168 }
        max_per_day: { type: integer, minimum: 1, maximum: 24, description: At most this many doses in any 24 hours }
        dose: { type: number, minimum: 0, description: The usual dose (fills the app's form) }
        dose_unit: { type: string }
        remind: { type: boolean, default: false, description: Notify phones when the next dose is due (server with notifications) }
    Reminder:
      type: object
      description: |
        Shown by the app when due; also sent to phones by a server with notifications set up.
        `sleep` counts time awake.
      required: [type, after_minutes]
      properties:
        type: { type: string, enum: [feed, sleep, diaper, pump] }
        after_minutes: { type: integer, minimum: 15, maximum: 1440 }
...
    SleepDetails:
      type: object
      required: [type]
      description: Needs `end` (use a sleep timer for a sleep in progress).
      properties:
        type: { type: string, enum: [sleep] }
        location: { type: string }
...
    ActivityDetails:
      type: object
      required: [type, kind]
      properties:
        type: { type: string, enum: [activity] }
        kind: { type: string, description: "free text, e.g. bath, tummy_time, outdoor, play, read, nail_trim, vitamin" }
...
    Event:
      oneOf:
        - $ref: "#/components/schemas/FeedEvent"
        - $ref: "#/components/schemas/SleepEvent"
        - $ref: "#/components/schemas/DiaperEvent"
        - $ref: "#/components/schemas/PumpEvent"
        - $ref: "#/components/schemas/GrowthEvent"
        - $ref: "#/components/schemas/HealthEvent"
        - $ref: "#/components/schemas/ActivityEvent"
        - $ref: "#/components/schemas/MilestoneEvent"
        - $ref: "#/components/schemas/NoteEvent"
      discriminator:
        propertyName: type
        mapping:
          feed: "#/components/schemas/FeedEvent"
          sleep: "#/components/schemas/SleepEvent"
          diaper: "#/components/schemas/DiaperEvent"
          pump: "#/components/schemas/PumpEvent"
          growth: "#/components/schemas/GrowthEvent"
          health: "#/components/schemas/HealthEvent"
          activity: "#/components/schemas/ActivityEvent"
          milestone: "#/components/schemas/MilestoneEvent"
          note: "#/components/schemas/NoteEvent"
    FeedEvent:
      allOf: [{ $ref: "#/components/schemas/EventCommonOutput" }, { $ref: "#/components/schemas/FeedDetails" }]
    SleepEvent:
      allOf: [{ $ref: "#/components/schemas/EventCommonOutput" }, { $ref: "#/components/schemas/SleepDetails" }]
    DiaperEvent:
      allOf: [{ $ref: "#/components/schemas/EventCommonOutput" }, { $ref: "#/components/schemas/DiaperDetails" }]
    PumpEvent:
      allOf: [{ $ref: "#/components/schemas/EventCommonOutput" }, { $ref: "#/components/schemas/PumpDetails" }]
    GrowthEvent:
      allOf: [{ $ref: "#/components/schemas/EventCommonOutput" }, { $ref: "#/components/schemas/GrowthDetails" }]
    HealthEvent:
      allOf: [{ $ref: "#/components/schemas/EventCommonOutput" }, { $ref: "#/components/schemas/HealthDetails" }]
    ActivityEvent:
      allOf: [{ $ref: "#/components/schemas/EventCommonOutput" }, { $ref: "#/components/schemas/ActivityDetails" }]
    MilestoneEvent:
      allOf: [{ $ref: "#/components/schemas/EventCommonOutput" }, { $ref: "#/components/schemas/MilestoneDetails" }]
    NoteEvent:
      allOf: [{ $ref: "#/components/schemas/EventCommonOutput" }, { $ref: "#/components/schemas/NoteDetails" }]
    DeletedEvent:
      type: object
      required: [id, child_id, deleted, updated_at]
      properties:
        id: { type: string }
        child_id: { type: string }
        deleted: { type: boolean, enum: [true] }
        updated_at: { type: string, format: date-time }
    EventOrDeleted:
      oneOf:
        - $ref: "#/components/schemas/Event"
        - $ref: "#/components/schemas/DeletedEvent"

    # -------- timers
    Segment:
      type: object
      required: [side, start, end]
      properties:
        side: { allOf: [{ $ref: "#/components/schemas/Side" }], nullable: true }
        start: { type: string, format: date-time }
        end: { type: string, format: date-time, nullable: true, description: null while this segment runs }
    Timer:
      type: object
      required: [id, child_id, kind, started_at, running, side, elapsed_seconds, segments, created_by, updated_at]
      properties:
        id: { type: string }
        child_id: { type: string }
        kind: { $ref: "#/components/schemas/TimerKind" }
        started_at: { type: string, format: date-time }
        running: { type: boolean, description: false while paused }
        side: { allOf: [{ $ref: "#/components/schemas/Side" }], nullable: true }
        elapsed_seconds: { type: integer, format: int64 }
        left_seconds: { type: integer, format: int64, description: breastfeed and pump only }
        right_seconds: { type: integer, format: int64, description: breastfeed and pump only }
        segments:
          type: array
          items: { $ref: "#/components/schemas/Segment" }
        created_by: { type: string, nullable: true }
        updated_at: { type: string, format: date-time }
...
    SyncEventIn:
      description: |
        A whole event as the device has it (`EventInput` fields; `start` is required here), plus
        `id`, `child_id` and `changed_at`; or a deletion with `deleted: true`.
      allOf:
        - type: object
          required: [id, child_id, changed_at]
          properties:
            id: { type: string, format: uuid }
            child_id: { type: string }
            changed_at: { $ref: "#/components/schemas/TimeInput" }
            deleted: { type: boolean, default: false }
          additionalProperties: true
    SyncTimerIn:
      type: object
      required: [id, child_id, kind, changed_at]
      properties:
        id: { type: string, format: uuid }
        child_id: { type: string }
        kind: { $ref: "#/components/schemas/TimerKind" }
        changed_at: { $ref: "#/components/schemas/TimeInput" }
        deleted: { type: boolean, default: false }
        segments:
          type: array
          items:
            type: object
            required: [start]
            properties:
              side: { $ref: "#/components/schemas/Side" }
              start: { $ref: "#/components/schemas/TimeInput" }
              end: { $ref: "#/components/schemas/TimeInput" }
    SyncResult:
      type: object
      required: [id, status]
      properties:
        id: { type: string }
        status: { type: string, enum: [applied, conflict, rejected] }
        message: { type: string }
        current: { description: "the server's copy, for conflict and rejected (null: none)" }

    # -------- import
...